Fix or silence some compiler warnings about narrowing conversions, unused variables and other minor stuff
(cherry picked from commit f64c555a25befd1ca585db1c0b7a8f9ca3d74ae1)

+++ test/test_arithmetic.hpp
@@ -812,14 +812,14 @@ void test_float_ops(const boost::mpl::int_<boost::multiprecision::number_kind_fl
       Real v(512);
       int exponent;
       Real r;
-      exponent = ilogb(v);
+      exponent = static_cast<int>(ilogb(v));
       r = scalbn(v, -exponent);
       BOOST_CHECK(r >= 1);
       BOOST_CHECK(r < std::numeric_limits<Real>::radix);
       BOOST_CHECK_EQUAL(exponent, logb(v));
       BOOST_CHECK_EQUAL(v, scalbn(r, exponent));
       v = 1 / v;
-      exponent = ilogb(v);
+      exponent = static_cast<int>(ilogb(v));
       r = scalbn(v, -exponent);
       BOOST_CHECK(r >= 1);
       BOOST_CHECK(r < std::numeric_limits<Real>::radix);
